requirements.txt
feedparser==6.0.11
requests==2.32.3
beautifulsoup4==4.12.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
html2text==2024.2.26
google-genai
supabase
fastapi
pydantic

